[REFACTOR] GSRecordControllerTest, GSRecordServiceTest import문 guest 경로 수정 (#116)
